[CI] Don't set test config for API microversions if master
If the version's being bumped on manila,
the newer API changes cause test failures with
manila-tempest-plugin.

We could carefully craft a
manila-tempest-plugin changeset that fixes
any incongruous test assumptions and make the
manila change depend on the tempest changeset.
However, that may make developing the API harder
than it is today.

Change-Id: Ie97fb49b38b565dcc57332db8b42605216372aa8
Related-Bug: #1781671