Improve share list speed using lazy='subquery'
lazy='immediate' leads to each relationship being collected when
it is accessed. This results in at least three extra queries when
we query for all share details. lazy='subquery' collects all data
when the query is exected. In this commit we only changed code for
improving the share list with details ("manila list") speed.

This is joint work together with Johannes Kulik.