Improve share list speed using lazy='subquery'
lazy='immediate' leads to each relationship being collected when
it is accessed. This results in at least three extra queries when
we query for all share details. lazy='subquery' collects all data
when the query is exected. In this commit we only changed code for
improving the share list with details ("manila list") speed.

Change-Id: Ia61b108ece0817069737980a614cc6c15c1a3507
Co-authored-by: Johannes Kulik <johannes.kulik@sap.com>