Introduce noop interface driver
In a Manila deployment using the Generic driver, it is possible
that the administrator has deployed manila-share on a machine that
already has access to the admin network used by Manila by being on
the same VLAN for example.

This means that it's not necessary for it to attempt to plug the ports
and do all the wiring due to the fact that it's already reachable,
therefore this patch introduces a new `interface_driver` which is
does nothing in order to avoid plugging/unplugging ports and avoiding
the requirement of running neutron-(openvswitch|linuxbridge)-agent
on the same node as manila-share.

The tests are pretty trivial because there's not much logic, but we
at least make sure that we don't call any ip_lib functions which means
we're not touching anything in host networking.

Change-Id: I2c3d6e4234bf7185b8da3c8e1701069c3a165ffc
(cherry picked from commit 6aec28f11f2ae7504d776b932800433263fd82da)