NetApp encrypt option differentiate between not present and false
If the encrypt extra spec is not set, now the setting is not treated
as 'false', i.e. no encryption. But the parameter will not be touched
to let the back end netapp operator decide. This allows for a mixed
setup of encrypted and non encrypted back ends without leaking this
detail up to the share type.

Change-Id: Ie76721cae7f48d006b25ddf5266d48136a3f7f40
(cherry picked from commit 8589e14e0c0f7781d1cf8f9c5b38cd9fa0708acb)