NetApp: Use cluster client in non-vserver contexts
When share server is in error state in Manila and backend vserver has
not been created, there exist entries of share server and shares in the
database. Deletion of a share in such case causes the share to end up
in error_deleting state. In order to delete the share correctly,
we need to fix _get_vserver which should check vserver existence with
a client that is not tunneled to the possibly not existing vserver.

Closes-Bug: #1964592
Change-Id: I1f6b94a7e54403a559a5d77b4d93b939a40e76bf