Allow disabling NFSv4.0
NFS v4.0 was previously handled differently in NetApp driver.
We want to explicitly disable v4.0 if it is not set in 'netapp_enabled_share_protocols'.

not sending the optional 'is-nfsv40-enabled' parameter to
'nfs-service-modify' means it will not be changed,
which is relevant for update_vserver in the ensure loop.

I.e. we don't want to disable nfs v4.0 for existing vservers.

Also enable v4.1 options like read/write delegation, pnfs, acls.

Change-Id: I1e2aa559cceede6505578bef38c1a5b2edf2d991