ccloud: Enable logical space reporting for new shares (#51)
This pr patches NetApp driver to set the Volume's space attributes `is_space_reporting_logical` and `is_space_enforcement_logical`. When a new Share is created, the attributes are set to True values. When a Volume is created based on an existing Volume, for example when creating a Share from Snapshot, or when creating a Share replica, the attributes are set to the same values as the source Volume.

Change-Id: I64bf5d915edac6784b114127958bcb612dd2ba48