ccloud: Disable cross-volume-deduplication to prevent metadata from filling volume (#52)
This PR patches the NetApp driver to disable cross volume deduplication
when necessary. When disabled, together with set sis state to
inline-only, we prevent the deduplication metadata from filling up the
Volume. When not disabled, the metadata could take capacity up to 4% of
the application's data in Volume. NetApp is working on a change to
OnTap, so that the metadata are stored seperated. When it is delivered,
we can remove this patch.