Allow resync if replica is out_of_sync
The 'snapmirrored' mirror-state indicates source and target volumes are
in snapmirror relation, but this does not mean that they are in sync, as
it is ASYNC mirror, there is always delta present which we have to take
care by doing updates/resyncs. So allow resync when replica is
out_of_sync take care of delta.