sapcc: try blocks for all promote additions
There is no rollback. We don't want to come to the situation again,
that we have an unfinished promote. It later got completely damaged by
an attempted repair in the periodic `update_replica_state`.

Also we had a case, where we don't know why the
`if effi_opts['cross_dedup_disabled']` codepath was taken with an
efficiency policy seemingly set to 'auto' (not 'inline-only') all the
time. Add logging for more insights the next time.

Change-Id: Iaa02525f70e557d98b33dc9fb5ce8c6afa440118