Soft delete volumes when they have clones (#116)
This PR allows user to delete the shares in Manila, even when it is not
allowed on the NetApp. The soft-deleted share is renamed and will be
pruned later by a periodic job in Manila.

(cherry picked from commit 5b62e751335d619200f467eb97d17f6f46a1d7c5)