NetApp: fix delay clone split when creating share from snapshot
When creating share from snapshot, the clone-split job is triggered
right after volume is created on NetApp. The subsequent tasks on the
volume might fail because of "Volume busy" error. To prevent this issue,
we would like to delay the clone split job to the end. This problem was
addressed before, but the previos fix does not work as expected. Both
split variable in arguments and split field in provisioning_options
should be considered, and split argument should take precedence.