Fix svm mig ipspace (#146)
This patch fixes the issue, that the destination share server is created
in the wrong ipspace if the corresponding vlan exists already on the
filer.  We make sure that the ipspace name is correctly determined from
configuration, rather than reusing the wrong ipspace name which existing
vlan attached to. The vlan is also moved to the right ipspace if
necessary.

Change-Id: I92d3b94955af798f1fe330e32bb319dbc931b559