sapcc: allow to ensure share instances in 'creating'
The list that is checked has only share instances that already
got a host field set, i.e. the creation stuck at a late stage
(after scheduling). Those should be recoverable here.

In the NetApp driver: always set the status to 'available'.
In case something goes wrong, the share instance is not returned
at all (and therefore not updated) in favor of setting status to
'error', which would be too invasive in certain situation.

And double check status in second ensure loop

in the meantime the share may have been deleted and
we don't want to update the status.

Especially with deferred deletion, shares are longer in that state,
where the backend share is still existing, i.e. the driver ensure
operations are working normally, but still we don't want to reset the
status to available.

And ensure that non busy shares have pools.

especially shares that ended up in extending_error due to not
having a pool should be fixed here.
And don't overwrite with potentially outdated host value
it may have been changed earlier in _ensure_share_instance_has_pool()

Change-Id: I53a65649ff1a589ce2c9ecec8ff0525f4d800925
Change-Id: Ie924e6fbe402a2ca1fb534bde178cd78a3eb5341
Change-Id: Ie949153e1bc2b854c4263f1899784b9dfe6ec094