manila-share: check service status
onyl setup the driver if the service is enabled.

Since the setup is wrapped in an infinite loop, it will pick up
if the service gets enabled and starts working.

The use-case is: I intentionally disabled a service and stopped it
(e.g. for doing maintenance on the back end).
Now I don't want it to be coming up if it gets started
(by someone that may be not aware of the maintenance)
unless it gets explicitly enabled again.

Change-Id: I58b377dbd49880761962a974a46f8b31b9b0af21