ccloud: force promote replica (#175)
This pull request introduces a force option to the share-replica-promote API.
Normally, Manila drivers should verify that data replication is in sync with the
source before promoting a replica to prevent data loss. Nevertheless, there are
situations where promotion may proceed despite the status of the replication.
The inclusion of this option in the API enables users to control the behavior of
the driver. Additionally, it enhances verification in the NetApp driver by
inspecting the health of SnapMirror before breaking it.

Change-Id: Ia08caf6d3f5864a112ce402567eddbd0d0aa1819