sapcc: double check status in second ensure loop
in the meantime the share may have been deleted and
we don't want to update the status.

Especially with deferred deletion, shares are longer in that state,
where the backend share is still existing, i.e. the driver ensure
operations are working normally, but still we don't want to reset the
status to available.

followup to 24fb551ee4098fc219bc31b77895b2a8ef463d66

Change-Id: Ie949153e1bc2b854c4263f1899784b9dfe6ec094