sapcc: keep lifs on vserver create error
don't delete, but disable.
followup to e43a45a4b5d7cdc53feef9a685ed61bc5d21f4c0

Caution: neutron ports are deleted anyhow to not block IPs.
That means in case of troubleshooting we need to make sure that
the lif IP that gets enabled again is free
But this makes it easier now, because we don't have to recreate the
lif.

Troubleshooting hints:
Maybe reserve it neutron again to avoid duplicate IP errors.
Check if the vlan is still valid, if no other vservers from the same
share network are on that filer, it may have been removed.
To force this: create a NFS share on that network on that filer
only then create the CIFS share you want to debug.

Change-Id: I208d0517bbc7898341faede81d1b2e6cdf3d4bb5