NetApp: start vserver delete by disabling the interfaces
Something may end up in error later, but we need to make sure that the
interfaces are no longer in use, since we already deleted the neutron
ports at that stage.
The interfaces will be deleted later implicitly.
Disabling is the safer approach, since this also works even if
the vserver still holds some orphaned volumes.

Change-Id: I65c285315850cac926e5439d17c144563215c45f