NetApp: start vserver delete by disabling the interfaces
Something may end up in error later, but we need to make sure that the
interfaces are no longer in use, since we already deleted the neutron
ports at that stage.
The interfaces will be deleted later implicitly.
Disabling is the safer approach, since this also works even if
the vserver still holds some orphaned volumes.

We don't need to disable already disabled interfaces, which would
happen on the second and consecutive runs of the cleanup if it failed.

Use cluster client, because if we turn off the vserver interface
the vserver api will not work ('13005:Unable to find API').

Change-Id: Ia34809d96097d684113e651e09676b4e8a643054