Add OpenAPI generation job
We want to start adding jsonschemas to the codebase of Manila in order
to be able to have a complete OpenAPI spec generated from the sources.
This is not always easy to judge whether the schema is looking good or
not (when corresponding validations are missing) therefore call a job
from codegenerator that generates OpenAPI. This gives possibility of
seeing how the spec is looking like and also visually compare result
with the expectation. Next step would be to add another job trying to
render API-Ref html docs from such an OpenAPI spec. Once all the
required information lands in the sources we could replace human written
API-Ref with the one generated from OpenAPI.

Change-Id: Ib7b99ed85541b9eb965f7a9941eadeffd8c7f3d9