prevent new vserver using cluster ipspaces
Orphan vlans are moved to DEFAULT ipspace when their originally
associated ipspaces are deleted. As a result, new vserver will be put in
the DEFAULT ipspace if it is on an orphan vlan. This PR will ensure a
private ipspace is created and used for vserver. The vlans are then
moved to the correct ipspace in the following steps.

Change-Id: I285a0e39603b6c144b3b8c8bafe61a48c420899d
(cherry picked from commit 76b32622f679a6f928a0c876cb02558c6df7dc5f)