cifs: enable SMB signing
by default it would be disabled.
It can only be enabled after cifs server has been created.
Therefore this can not be coupled with the other cifs security settings
that are handled in configure_cifs_encryption() before cifs server
creation. We wait 5 minutes for cifs server creation to finish.
If this does not finish in time, we proceed anyway, since cifs setup is
already very prone to errors.

equivalent to ontap client
'vserver cifs security modify -is-signing-required true [...]'

We have it enabled everywhere, so we don't need to ensure this.
On demand we may even disable again(?).

If we want to enable ensuring again, the code should be improved to
first check if it is already enabled, and send the enable call only
if necessary.

Change-Id: I3f1bbf009f47ebcb00c4c19206e08f11a87da9f3