Lock shares when creating a locked access rule
Locks for access rules currently only lock the access rule, but
in case someone requests the share to be deleted, we will proceed
with the requests in case it doesn't have any locks.

Now, when creating access rules and locking them against deletion,
shares will also be locked to avoid their deletion and getting
clients disconnected.

For existing access rules, to avoid the share being deleted by
someone else, a share lock must be created.

The share locks will use a specific lock reason which will be tied
to the UUID of the access lock, and we will delete the share logs
alongside the access rules locks together.

Closes-Bug: #2075967
Change-Id: Ie5204e6098a9338ea721207e077af117793eb255
Signed-off-by: Carlos Eduardo <ces.eduardo98@gmail.com>