modify lazy loading for ShareInstance model
The ShareInstance model has relationships with AvailabilityZone and
ShareType models. These relationships are loaded lazily by subquery.
This emits sql queries for availability zone and share type with
subquery for all share instances, which is not efficient as the share
instance table grows. This change modifies the relationships to load the
availability zone and share type with selectinload, which will load all
items from the availability zone and share type tables in a single
query. It's acceptable because they are pretty small tables.

Change-Id: I0525b8cff6616e4e6f290de50dcdd50e1b7435d8