cifs: don't ensure smb signing
follow up to 688d856f31597ff27f678df6452e2c53aa4008eb

Now we have it enabled everywhere, we don't need to ensure this.
On demand we may even disable again(?).

If we want to enable this again, the code should be improved to
first check if it is already enabled, and send the enable call only
if necessary.

Change-Id: I40533ec3f40e53f06098260e2290247dd386878e