[NetApp] Fixed share creation workflow via REST
At present, in share creation workflow when dedupe and compression are not enabled, it is trying to disable dedupe first and ended up with the error mentioned in the bug.

As part of this patch, we are changing the order to disable compression first followed by disabling dedupe so that the share creation workflow goes through.

PS: ONTAP compression requires dedup to be enabled.

Closes-Bug: #2080951
Change-Id: I2461d232060c996e8a4a2bd8d96a1327040c82f1