db purge_deleted_records: make sure that entries are soft-deleted
In addition to solely relying on `deleted_at` column set, the purge
method now also checks the `deleted` column to safeguard only purging
records that really have been soft-deleted.

Outcome of an incident, where a soft-deleted record had been
restored/undeleted, but without touching the deleted_at field.
The idea was to not touch deleted_at at that time to have some history
on that previous delete. Now this history is only available for share
instances with the `terminated_at` column.

Also fix the test to work on soft-deleted objects. It only
operated on objects that had deleted_at set beforehand.

Change-Id: I23105561a2d7cc65d99da7f4d39d51886ea15a26