sapcc: fix rest api force delete
The force option needs to be a query param.
And since we don't have proper error handling for rest api in this
release, we use the timeout query to not have to deal with async jobs.

Now we apply force delete properly for 2-hour old shares instead of
just for 2-minute old ones.

Change-Id: Id2fd93d7d6f4d53d788a40243d174fe96be49def