sapcc: improve error handling on prune deleted volumes (#228)
sapcc: improve error handling on prune deleted volumes

don't make the loop stop on error, we want to try pruning again in the
next run.

The prune deleted volumes was operating on deleted volumes (type DEL)
that were already in the recovery queue. Those should not be touched,
they don't need to be brought back online to solve clone split
situations. We are only interested in type RW. So we also don't do
housekeeping on e.g. DP volumes.

Anyway we also catch errors on split start now for a smoother
continuation of the loop.

Change-Id: I7a3c83423d731c4922c8f0f570fa9309fc70db46