fix race in share network delete
We iterate over the share network subnets, send rpc calls to delete
their share servers and then remove the share network subnet and the
share network from the db by using soft delete.

The info from those models is still required to deallocate the
network in the driver, e.g. the neutron plugin needs the neutron
net info.

We don't want the API layer of share network delete to be blocked
waiting for share server deletion, instead we look at the soft deleted
data in the manager handling the share server deletion.
This usually happens shortly after the soft deletion was done, so we
should still be able to find the entries.

Change-Id: I1f23b9e93316cabb7bd32be7bb9b63dbd4eb889f