NetApp share server migration: re use existing ipspace
It works currently, if the existing ipspace name has the same name
that we would try to create, because create_ipspace in the client
accepts that.

But if the existing ipspace name is different, because we have
multiple neutron subnets, server migration fails.

In the future, once all ipspaces are based on neutron network name
instead of subnet name, this will be needed only on edge cases.

Change-Id: If220727ded0bc796645eb29852d18c23e0a9e8b2