NetApp disable nfs v4.0 for new share servers
this has been the default for ONTAP 9.7 and 9.8 and we mistakenly
relied on that it would stay that way in
3503a199a00cef176c398b98a4187808f5d521ad

But since ONTAP 9.9.1 the default is to enable nfsv40 again.

Now making the disablement explicit, since we now have a proper way to
distinguish between options that should be set on create and those
that should be set on update, too.

Change-Id: I39e2da2c3393ec95312c09a608a997a29981a471