Dummy db migration for backporting
This PR includes DB migration skipped when backporting backup features
to antelope. The files are just place holder with version id. The real
migration code is commented out.

To upgrade, we need to do the opposite: include the real migration code
for the place holders, and comment out the migration code in backup
features.