sapcc: add request_id if we already have any in the header
This is only a quick-fix for the problem that the request id
changes in-between the same request.

The proper way would be to do a similar
refactoring like https://github.com/openstack/cinder/commit/18b8033b6bd1945bd27b03b779a32a883df5fb10
because doing this in the auth middleware here is actually too late
in the stack.

Change-Id: I4183757e02c7c6446911ace898499af303d711e5