db: improve share get with filters and limit results
When applying both a filter and a pagination limit to share get
queries, the resultset may contain fewer shares then expected,
because the limit is applied to the set of shares joined by their
share instances.

This only avoids getting wrong number of shares in the result.
If the filter was applied to any proxified property of share
instances, the result may still contain non-matching elements.

A follow-up patch for this is needed to maybe replicate the behaviour
of the `instance` property of the Share db model into the
`_share_get_all_with_filters()` method

Change-Id: Ic22c556c96fd097e915dd1a028d43c5a298bac17
(cherry picked from commit 8603065887d5c953ab1adfa6de056a1885da4df5)