[NetApp] fix resetting autosize attributes while replica promotion
Instead of calling NetApp API via the modify_volume() function, create a
new function in the NetApp client to reset the autosize attributes
alone. The former touches various attributes, such as `snapshot-policy`,
which are not allowed to be modified in the stage of promoting replica.

Todo: Submit this change can to upstream.
Change-Id: Ie69b08c8555d42d425db86b9ef55c76fe50b0376