Allow list values for same_host/different_host hint
When setting the scheduler hints same_host/different_host, upstream
allows only string values. However, we had a patch that allows list
values as well. The patch was hidden in the giant squash commit during
Xena upgrade. We need to bring this change back when upgrading to newer
release to stay consistent with previous behavior.

Change-Id: If63f2755b8c6d4a059921f4d13e4f687c8b2b2a2